Move package for classes `CopyMessage` and `ForwardMessage`
